Slide Creation
This is my first attempt of getting all of the written material into slide cells.
On future revisions I hope to make the slides look more digestible and to make them flow better.

Additionally, I need to still add the images,regression results, and revise the results section of the paper.

While I did not use ChatGPT for this revision, I plan to do so on my next attempt to help with condensing the slides.