Slide Update
Attempted to create the remaining slides and add images. This was my first attempt adding images and failed to do so.

I reference git hub sources and Chat GPT with the following prompt but found little to no success.

ChatGPT Prompt: "You are a very gifted coder with intense python and jupyterbook knowledge. How do you add an image to a slide within jupyterbook."

The response was the code that I used for this attempt but my slides still fail to produce the image.

During my next iteration I will try to further solve this problem and begin making the slides more condense.